Working on reservation adding. TODO : move try-catch to DatabaseConnection

+++ src/main/java/data/Movie.java
@@ -2,15 +2,15 @@ package data;
 
 public class Movie {
 	
-	private final long id;
+	private final int id;
 	private final String title;
 	
-	public Movie(long id, String title) {
+	public Movie(int id, String title) {
 		this.id = id;
 		this.title = title;
 	}
 	
-	public long getId() {
+	public int getId() {
 		return id;
 	}
